REFACTOR: BB-1521 Move more code from PivotTable
